renames camel case file to kebab spec
