+++ src/server/main.js
@@ -24,7 +24,7 @@ const getProjectManager = require('../aws/project-init');
 const daemons = require('./daemons');
 const pushHandler = require('./push');
 const authentication = require('./auth/authentication');
-const githubAuthMiddleware = require('./auth/githubHook');
+const githubAuthMiddleware = require('./auth/github-hook');
 const users = require('./auth/users');
 const util = require('../util');
 const clusternatorApi = require('./clusternator-api');
